Update Jenkinsfile (#138)
In the pipeline library I updated the go binary to add a comment to a PR when Jenkins is running. This will be used by Manager in the future. Unfortunately this means that nothing can run on master, since master does not have go installed. The mainFlow() was updated accordingly but Manager Jenkins did not use it. This change fixes that.