bugfix: remove istio initializer from istio-system workloads
istio-system namespace is handled differently than normal user
namespaces and should not have sidecar's injected. This fixes a bug in
that logic where the sidecar was not injected (correct) but the
pending list of initializers was not cleared (bug). This caused
istio-system components to not be deployed in cluster-wide
installations to istio-system name.