Port of Pilot to CentOS
CentOS base OS requires iptables from CentOS containers and
its library dependencies, not from Ubuntu.  One possible solution
to this problem is to statically build iptables which may fix the
problem and allow one consistent set of images to be used for all
of the pilot repo.  I'll explore that shortly via COPR and see if
I can build a static xbtables.