Distribute new protobuf representation of mixerclient configuration (#1644)
* 1) generated mixerclient protobuf code. This will likely be simplified
(or go away) once we start storing generated artifacts as part of repo
consolidation.

2) Distribute the new protobuf mixerclient configuration as part of
the existing mixer filter configuration. The protobuf is encoded as a
JSON map value under "v2" key in the filter config struct map
alongside the legacy ad-hoc configuration. This is a backwards
compatible change to allow proxies to incrementally adopt the new
configuration. The legacy ad-hoc config is deprecated and will
eventually be removed.

3) annotate inbound route's opaque_config with the destination.service
of the server instance. Mixerclient can use this to look-up
per-service quota/api in the filter configuration.

* Address review comments and add per-service control config

* encode IP attributes as bytes

* create per-service control config and add destination.service to the
  list of attributes to include in check/report calls.

* add a warning of v2 config encoding fails

* fix warning

* refresh ISTIO_API sha

* create ControlConfigs map and refresh test data

* add comment